+++ src/main/kotlin/com/github/rafaelldi/tyeplugin/services/TyeGlobalToolService.kt
@@ -1,6 +1,6 @@
 package com.github.rafaelldi.tyeplugin.services
 
-import com.github.rafaelldi.tyeplugin.actions.InstallTyeGlobalToolNotificationAction
+import com.github.rafaelldi.tyeplugin.actions.notification.InstallTyeGlobalToolAction
 import com.github.rafaelldi.tyeplugin.settings.TyeSettings
 import com.github.rafaelldi.tyeplugin.util.ToolVersion
 import com.intellij.execution.configurations.GeneralCommandLine
@@ -74,7 +74,7 @@ class TyeGlobalToolService(private val project: Project) {
         val isTyeGlobalToolInstalled = isTyeGlobalToolInstalled()
         if (!isTyeGlobalToolInstalled) {
             Notification("Tye", "Tye is not installed", "", NotificationType.WARNING)
-                .addAction(InstallTyeGlobalToolNotificationAction())
+                .addAction(InstallTyeGlobalToolAction())
                 .notify(project)
             return
         }
@@ -107,6 +107,35 @@ class TyeGlobalToolService(private val project: Project) {
                         BrowserUtil.browse("https://aka.ms/failure-installing-tool")
                     }
                 })
+                .notify(project)
+        }
+    }
+
+    fun uninstallTyeGlobalTool() {
+        val isTyeGlobalToolInstalled = isTyeGlobalToolInstalled()
+        if (!isTyeGlobalToolInstalled) {
+            Notification("Tye", "Tye is not installed", "", NotificationType.WARNING)
+                .addAction(InstallTyeGlobalToolAction())
+                .notify(project)
+            return
+        }
+
+        val commandLine = GeneralCommandLine()
+            .withParentEnvironmentType(GeneralCommandLine.ParentEnvironmentType.CONSOLE)
+            .withExePath("dotnet")
+            .withParameters("tool", "uninstall", "Microsoft.Tye", "--global")
+        val output = ExecUtil.execAndGetOutput(commandLine)
+        val success = output.exitCode == 0
+
+        if (success) {
+            log.info("Tye is successfully uninstalled")
+
+            Notification("Tye", "Tye is successfully uninstalled", "", NotificationType.INFORMATION)
+                .notify(project)
+        } else {
+            log.error(output.stderr)
+
+            Notification("Tye", "Tye uninstallation failed", output.stderr, NotificationType.ERROR)
                 .notify(project)
         }
     }